Move source code to src folder Such move should allow creation of separate folder for tests
